presidents/requirements.txt
Django==1.11.3
nltk==3.2.1
Markdown
django-filter
djangorestframework
pandas
pillow
requests
beautifulsoup4
numpy
scipy
scikit-learn

